VendingMachine refactored into separate files
