+++ Wrappers.cpp
@@ -11,3 +11,15 @@ CatWrapper::~CatWrapper()
     delete catPtr;
     std::cout << "/// CatWrapper ///" << std::endl;
 }
+
+VendingMachineWrapper::VendingMachineWrapper(VendingMachine* ptrToVendingMachine) : vendingMachinePtr(ptrToVendingMachine)
+{
+    std::cout << "+++ VendingMachineWrapper +++" << std::endl;
+}
+
+VendingMachineWrapper::~VendingMachineWrapper()
+{
+    delete vendingMachinePtr;
+    std::cout << "/// VendingMachineWrapper ///" << std::endl;
+}
+
