+++ Wrappers.h
@@ -1,9 +1,17 @@
 #pragma once
 #include "Cat.h"
+#include "VendingMachine.h"
 
 struct CatWrapper
 {
     Cat* catPtr = nullptr;
     CatWrapper(Cat* ptrToCat); 
     ~CatWrapper();
+};
+
+struct VendingMachineWrapper
+{
+    VendingMachine* vendingMachinePtr = nullptr;
+    VendingMachineWrapper(VendingMachine* ptrToVendingMachine);
+    ~VendingMachineWrapper();
 };
